split indexes client
